Add files via upload
Q1. Create a Pandas Series that contains the following data: 4, 8, 15, 16, 23, and 42. Then, print the series.

Q2. Create a variable of list type containing 10 elements in it, and apply pandas.Series function on the
variable print it.

Q3. Create a Pandas DataFrame that contains the following data:
Then, print the DataFrame.

Q4. What is ‘DataFrame’ in pandas and how is it different from pandas.series? Explain with an example.
Q5. What are some common functions you can use to manipulate data in a Pandas DataFrame? Can
you give an example of when you might use one of these functions?

Q6. Which of the following is mutable in nature Series, DataFrame, Panel?

Q7. Create a DataFrame using multiple Series. Explain with an example.